requirements.txt
asyncio==3.4.3
sanic==0.5.4
uvloop==0.8.0
raven==6.1.0
kafka-python==1.3.4
requests==2.18.4